remove intercepted routes and use agent detail page (with modal looking)
