--- web-app/src/app/(landing)/@agentDetail/default.tsx
@@ -1,3 +1,0 @@
-export default function Default() {
-  return null;
-}
