--- web-app/src/app/app/@agentDetail/default.tsx
@@ -1,3 +1,0 @@
-export default function Default() {
-  return null;
-}
