+++ web-app/src/components/agents/agent-detail-link.tsx
@@ -1,7 +1,4 @@
-"use client";
-
 import Link from "next/link";
-import { usePathname } from "next/navigation";
 
 interface AgentDetailLinkProps {
   children: React.ReactNode;
@@ -14,16 +11,6 @@ function AgentDetailLink({
   agentId,
   className,
 }: AgentDetailLinkProps) {
-  const pathname = usePathname();
-
-  if (pathname.startsWith("/app")) {
-    return (
-      <Link className={className} href={`/app/agents/${agentId}`}>
-        {children}
-      </Link>
-    );
-  }
-
   return (
     <Link className={className} href={`/agents/${agentId}`}>
       {children}
