+++ web-app/src/components/agents/agent-detail/card-section.tsx
