+++ web-app/src/components/agents/agent-detail/index.ts
@@ -1,3 +1,5 @@
+export * from "./agent-detail";
+export * from "./card-section";
 export * from "./section-1";
 export * from "./section-2";
 export * from "./section-3";
