+++ web-app/src/components/agents/agent-detail/section-1.tsx
