+++ web-app/src/components/agents/agent-detail/section-2.tsx
