+++ web-app/src/components/agents/agent-detail/section-3.tsx
