+++ web-app/src/components/agents/agent-detail/section-4.tsx
