+++ web-app/src/components/agents/agent-detail/section-5.tsx
