--- web-app/src/components/agents/agent-modal/card-sections/card-section-3.tsx
@@ -1,25 +1,0 @@
-import {
-  AgentDetailSection3,
-  AgentDetailSection3Skeleton,
-} from "@/components/agents/details";
-import { AgentWithRelations } from "@/lib/db";
-
-import { CardSection } from "./card-section";
-
-function CardSection3({ agent }: { agent: AgentWithRelations }) {
-  return (
-    <CardSection>
-      <AgentDetailSection3 agent={agent} />
-    </CardSection>
-  );
-}
-
-function CardSection3Skeleton() {
-  return (
-    <CardSection>
-      <AgentDetailSection3Skeleton />
-    </CardSection>
-  );
-}
-
-export { CardSection3, CardSection3Skeleton };
