--- web-app/src/components/agents/agent-modal/card-sections/index.ts
@@ -1,6 +1,0 @@
-export * from "./card-section";
-export * from "./card-section-1";
-export * from "./card-section-2";
-export * from "./card-section-3";
-export * from "./card-section-4";
-export * from "./card-section-5";
