--- web-app/src/components/agents/agent-modal/index.ts
@@ -1,2 +1,0 @@
-export * from "./agent-modal";
-export * from "./agent-modal-content";
