+++ web-app/src/components/agents/index.ts
@@ -1,6 +1,7 @@
 export * from "./agent-badge-cloud";
 export * from "./agent-bookmark-button";
 export * from "./agent-card";
+export * from "./agent-detail";
 export * from "./agent-detail-link";
 export * from "./agent-hire-button";
 export * from "./agent-modal";
